Refactor: subdivision entre modulo, cobertura con unit test en cada file
