+++ Cargo.toml
@@ -19,5 +19,10 @@ predicates = "3.1"
 toml = "0.9.4"       
 serde = "1.0"      
 
+[lib]
+path = "src/lib.rs"
 
+[[bin]]
+name = "todo"
+path = "src/main.rs"
